requirements.txt
# PanelFM: Time Series Foundation Models for PMPM Cost Prediction

# Core
numpy>=1.24
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3

# Gradient Boosting
xgboost>=2.0
lightgbm>=4.0

# Time Series Foundation Models
chronos-forecasting>=1.4
torch>=2.1

# Time Series Baselines
statsforecast>=1.7

# Visualization
matplotlib>=3.8
seaborn>=0.13

# Data
pyarrow>=14.0
pyyaml>=6.0

# Utilities
tqdm>=4.66
joblib>=1.3
